v1.0.0-221111
- Atualização da Documentação
- Inclusão de Dockerfile e instruções para implantação
- Criação de função utilitária para compor queries
- Organização de importação de funções de serviço e funções utilitárias
