--- CHANGELOG.md
@@ -1,47 +1,0 @@
-<hr>
-<h1 align="center">changelog</h1>
-<p align=center><i align="center">Notas de versão</i></p>
-
-<hr>
-
-<div align="center">
-<a href="">[![Last Commit](https://img.shields.io/github/last-commit/frtechdev/flem-bd-legado-api)](https://github.com/frtechdev/flem-bd-legado-api/) </a>
-<br>
-
-</div>
-
-## v0.0.2-220721dx
-
-- Alteração da localização da documentação feita pelo jsdoc
-- Atualização da documentação
-- Alteração dos parâmetros de redirecionamento da página principal da API para redirecionar para a documentação
-- Documentação dos métodos e demais componentes
-- Definição da Homepage da API
-- Remoção da especificação de porta via .env (utilizando sintaxe direta dos scripts de execução do Node)
-- Atualização do README
-- Reorganização dos utilitários e ferramentas de serviço
-- Criação de controller para listagem e verificação de Beneficiários
-- Criação de controller para listagem e verificação de Situações
-- Criação de controller para listagem e verificação de Vagas
-- Alterado algoritmo para manipulação das requisições API aos Controllers para trabalhar com filtros e critérios específicos de acesso ao BD
-- Criado utilitário para realizar parsing dos objetos das requisições API e montar os filtros
-
-## v0.0.1-220701dx
-
-- Criação dos arquivos principais do projeto
-- Criação e definição das variáveis de ambiente
-- Criação de linting do ES para Next
-- Criação do gitattributes
-- Criação do gitignore
-- Adição de regras de linting para markdown
-- Criação do changelog
-- Criação do arquivo de Schema do ORM
-- Atualização do README
-- Criação de rota para fornecer dados de beneficiários
-- Criação de utilitários para formatação de CPF, Data e Nomes Próprios (capitalização)
-
-<br>
-
-## Initial Commit
-
-- Initial Commit
