+++ src/services/index.js
@@ -1,0 +1,2 @@
+export { allowCors } from "./apiAllowCors";
+export { prisma,queryComposer } from "./prisma";
