+++ src/utils/index.js
@@ -1,0 +1,6 @@
+export { maskCapitalize, maskCPF, maskDate, unmaskCPF } from "./masks";
+export {
+  parseArrayToInteger,
+  parseArrayToString,
+  parseArrayToStringEquals,
+} from "./parsers";
